requirements.txt
Django==2.0.5
djangorestframework==3.8.2
Pillow==5.2.0
requests==2.19.1
